Move gallery chapter interfaces to models
